Manage resources UI made to be User friendly
Instead of duplicate field & text & ... for adding and subtracting balance it all converted to be one and rather having a button to change the state of being addition or subtracted.

Add dialogbox to ask for user's confirmation before adding or subtractin balance for error-prone.

States will be stored when changing form portrit mode to landskape and vice-versa.

Plus and Minus icons used in this section of app for chaning the Balance type:
https://www.iconfinder.com/icons/126583/add_plus_icon#size=20
https://www.iconfinder.com/icons/183199/minus_icon#size=20